VA: dialer-less HTTP-01 rewrite.
Continued bugs from the custom dialer approach used by the VA for
HTTP-01 motivated a rewrite.

Instead of using a custom dialer to be able to control DNS resolution
for HTTP validation requests we can construct URLs for the IP addresses
we resolve and overload the Host header. This avoids needing the dialer
but still prevents Go from using the system resolver.