Remove NewCertificatesTick from ocsp-updater.
Since https://github.com/letsencrypt/boulder/pull/2633 we generate OCSP
at first issuance, so we no longer need this loop to check for new
certificates that need OCSP status generated. Since the associate SQL
query is slow, we should just turn it off.

Also remove the configuration fields for the MissingSCTTick. The code
for that was already deleted.