Break dependency on Golang docker image (#4028)
When a Golang security release comes out, we don't want to wait on
the upstream Docker image to update and upload before we can start
running tests on it. This changes our boulder-tools image so it
can download and install Golang itself.

Also, fix some issues in build.sh:

Unparallelize the `go get` with the `apt install`. The `go get`
pulls in `go-sqlite3`, which needs `cmake`, installed by `apt install`.
    
Also, don't remove build-essential and cmake at the end.