SA: Set issuedNames renewal bit in AddCertificate.
The `renewal` field of the `issuedNames` table is indexed to allow more
efficient processing of rate limit decisions w.r.t. renewals for the
certificates per domain rate limit. Before we can start using this field
in rate limit calculations we need to populate it appropriately. A new
`SetIssuedNamesRenewalBit` feature flag for the SA controls whether we
do so or not.