SA: refactor DB stat collection & collect more stats.
Go 1.11+ updated the `sql.DBStats` struct with new fields that are of
interest to us.

The existing autoprom stats code is replaced with new first-class
Prometheus metrics.

The `max_db_connections` stat from the SA is removed because the Go
1.11+ `sql.DBStats.MaxOpenConnections` field will give us a better view
of the same information.

The autoprom "reused_authz" stat that was being incremented in
`SA.GetPendingAuthorization` was also removed. It wasn't doing what it
says it was: counting reused authorizations and was instead counting the
number of times `GetPendingAuthorization` returned an authz.