ci: restore load-generator run.
This restores running the `load-generator` during CI to make sure it
doesn't bitrot. It was previously removed while we debugged the VA
getting jammed up and not cleanly shutting down.

Since the global `pebble-challtestsrv` and the `load-generator`'s
internal chall test srv will conflict this requires moving the
`load-generator` run to the end of integration tests and updating
`startservers.py` to allow the load gen integration test code to stop
the `pebble-challtestsrv` before starting the `load-generator`.

The `load-generator` and associated config are updated to allow
specifying bind addresses for the DNS interface of the internal
challtestsrv. Multiple addresses are supported so that the
`load-generator`'s chall test srv can listen on port DNS ports Boulder
is configured to use. The `load-generator` config accepts a `fakeDNS`
parameter that can be used to specify the default IPv4 address used.