va: add specific error for HTTP-01 to HTTP/2 server.
In practice the only way to add a specific error for this case is with
a regex on the error string.

The error returned from the stdlib `http.Client` for HTTP to an HTTP/2
server is just an `errors.ErrorString` instance without any context
(once you peel it out of the wrapping `url.Error`).

Even directly in the stdlib code at the place in `http/response.go` that
generates the error it's using a `&badStringError` and just putting the
byte string that `textproto` read into it.

To detetc this case in `detailedError` I added a pre-compiled regex that
will match the net/http malformed HTTP response error with the raw bytes
matching an arbitrarily sized HTTP/2 SETTINGS frame. Per RFC "A SETTINGS
frame MUST be sent by both endpoints at the start of a connection" and
so this seems like a fairly reliable indicator of an unexpected HTTP/2
response in an HTTP/1.1 context.