PA: Support YAML for hostname policy.
Our existing policy configs use JSON. We would like to switch to YAML to
match the rate limit policy configs and to support commenting/tagging
entries in the hostname policy.

The PA is updated to support both JSON and YAML while we migrate the
existing policy data to the new format. To verify there are no changes
in functionality the existing unit tests were updated to test the same
policy expressed in YAML and JSON form.

In integration tests the `config` tree continues to use a JSON hostname
policy file to test the legacy support. In `config-next` a YAML hostname
policy file is used instead.