admin-revoker: use authz2 SA revocation RPC.
The `RevokeAuthorizationsByDomain` SA RPC is deprecated and
`RevokeAuthorizationsByDomain2` should be used in its place.

NOTE: The new `RevokeAuthorizationsByDomain2` RPC doesn't return the
number of pending/final authorizations that were revoked. Because of
this the output for the admin-revoker has changed to only confirm
revocation took place and not list how many authzs were revoked.