Run `go mod tidy`.
This introduces several indirect dependencies that we didn't have
before. This is because `godep` used to operate at a package level, but
`go mod` operates at a module (~= repository) level. So if we used a
given repository, but didn't use all of its packages, we wouldn't
previously care about the transitive dependencies of the packages we
weren't using. However, in the `go mod` world, once we care about the
repository, we care about all of that repository's transitive
dependencies.