publisher: disable HTTP/2 support.
In Boulder Issue 3821[0] we found that HTTP/2 support was causing hard
to diagnose intermittent freezes in CT submission. Disabling HTTP/2 with
an environment variable resolved the freezes but is not a stable fix.

Per the Go `http` package docs we can make this change persistent by
changing the `http.Transport` config:

  Programs that must disable HTTP/2 can do so by setting
  Transport.TLSNextProto (for clients) or Server.TLSNextProto (for
  servers) to a non-nil, empty map"

[0]: https://github.com/letsencrypt/boulder/issues/3821