RA: fix suberror identifier bug in CAA recheck.
In the RA's `checkAuthorizationsCAA` function we loop through a list of
`*core.Authorization`s, dispatching each to a Go routine that checked
CAA for the authz and wrote an `error` to a results channel.

Later, we iterate the same `*core.Authorization` list and read `error`s
from the channel. If we get a non-nil error, then the current
iteration's `*core.Authorization` is used as the identifier for the
suberror created with the non-nil error.

This is a flawed approach and relies on the scheduling of goroutines
matching the iteration of the authorizations. When the goroutines write
results to the channel out of order the loop over the authorizations
will construct a sub-error with the wrong identifier. This manifests as
making the `TestRecheckCAAFail` unit test appear flaky, because it
specifically checks the expected identifiers in the returned
subproblems.

The fix involves writing both the checked authorization and the error
result to the results channel. Later instead of iterating the
authorizations we just read the correct number of results from the
channel and use the attached authorization from the result when
constructing a sub-error