test: remove stale comment in sa_db_users.sql
I noticed this stale bit of text in `test/sa_db_users.sql` while working
on something unrelated.

The CA component does not have its own DB, it relies on the SA and its
DB the same as other components.

Similarly we don't need to drop each user and recreate them, we use
`CREATE USER IF NOT EXISTS` from MariaDB 10.1+