va: don't wrap up http-01 DNS problem types.
When `va.newHTTPValidationTarget` gets an error from `va.getAddrs` it
does not need to wrap the error in a `berrors.ConnectionFailureError`.
The `getAddrs` function will only ever return a `probs.DNS` or
a `probs.UnknownHost` result and the va `detailedError` function can
handle these fine as-is. This is the approach taken by the TLS-ALPN-01
validation code already.