va: clean up DNS error handling for HTTP-01 challenges.
The `va.getAddrs` function is internal to the VA and can return
`berrors.BoulderError`s with a DNS type when there is an error,
allowing the calling code to convert this to a problem when required
using the existing `detailedError` function. This avoids some clunky
conversion the HTTP-01 code was doing that misrepresented DNS level
errors as connection problems with a DNS detail message.

In order to add an integration test for challenge validation that
results in `getAddrs` DNS level errors the Boulder tools image had to be
bumped to a tag that includes the latest `pebble-challtestsrv` that
supports mocking SERVFAILs. It isn't possible to mock this case with
internal IP addresses because our VA test configuration does not filter
internal addresses to support the testing context.