orphan-finder: fix OCSP service config field
In d3b9107 the orphan-finder gained the ability to generate OCSP
responses. To support this the `config-next/orphan-finder.json` config
was updated with a `ocspGeneratorService` key.

This update changes the orphan-finder's `config` struct to match the
JSON key in-use. Prior to this change the struct field was expecting to
be matched to a `caService` key in the JSON config.