Error messages made more intuitive
LE is popular and aims to popularise certificate issuance. End users who see error messages cannot be assumed to be as DNS-experienced as previously. The error messages in this file are terse and unobvious to the point that they are often unlikely to be well understood by those they are intended to inform, who may be "just trying to get a LE cert for their domain". 

- Example 1: the rightmost label is called a "TLD" (widely understood)  in one error message but a "public suffix" (not widely understood term) in another.
- Example 2: The "DNS name" is a fairly well understood term (although simply "domain name" is very clear to most). But "label" is not a well known term at all, and may even be perceived as some other item of data.
- Example 3: References are made to unacceptable lengths and sizes, but the user isn't informed of the acceptable ones where known, which allows correction of the error.

This PR suggests improvements that may help LE users.