Add CheckForACMEV1Authz flag.
This flag defaults to true (our current behavior). If we disable it, it
removes the check in GetAuthorizations2 that ensures we don't reuse
ACMEv1 authorizations in an ACMEv2 order. That check can be an expensive
database query in certain situations, so we may want to disable it
temporarily while we implement a faster query.