bdns: move logDNSError to exchangeOne, log ErrId specially.
We've found we need the context offered from logging the error closer to
when it happens in the `bdns` package rather than in the `va`. Adopting
the function requires adapting it slightly. We know it won't be called
with any timeout results, with a non-dns error, or with a nil underlying
error.

Having the logging done in `bdns` (and specifically from `exchangeOne`)
also lets us log the wire format of the query and response when we get
a `dns.ErrId` error indicating a query/response ID mismatch. A small
unit test is included that ensures the logging happens as expected.