SA: conditionally track issued names in AddPrecertificate.
Prev. we inserted data for tracking issued names into the `issuedNames`
and `fqdnSet` tables during `sa.AddCertificate`. A more robust solution
is to do this during `sa.AddPrecertificate` since this is when we've
truly committed to having issued for the names in the precert.

The new SA `WriteIssuedNamesPrecert` feature flag switches to writing
these tables during `AddPrecertificate` when enabled.